0.0.3

+++ tests/utils.py
@@ -1,15 +1,23 @@
 import random
 import string
 from falcon.testing import TestCase
+from redis import StrictRedis
+from blog.constants import BLOG_REDIS_HOST, BLOG_REDIS_PORT
 from blog.db import client
-from blog.middleware.redis import client as redis
 
 
-def drop_database():
-    """Drop all collections in database"""
+def drop_database(drop_redis: bool=True):
+    """
+    Drop all collections in database.
+
+    :param drop_redis: If toggled, will flush redis database.
+    :type drop_redis: bool
+    """
     db_name = client.get_database().name
     client.drop_database(db_name)
-    redis.flushdb()
+    if drop_redis:
+        redis = StrictRedis(host=BLOG_REDIS_HOST, port=BLOG_REDIS_PORT)
+        redis.flushdb()
 
 
 def random_string(length: int) -> str:
